[Refactoring] Using OneToOneField to extend models
Previously, I used connections that were described for every model
that I have, which means that, in order to change them, you would have
to change the code in various locations. That, of course is a nightmare
to maintain, so I refactored it.

The class ConnectionDetails holds information about connections, and
other classes which need that information simply refer to them via
OneToOneFields. You can find that class in a new "app" - "connection".

Very simillar thing could be said about PrivacyControls.